% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_scientific.R
\name{fancy_scientific}
\alias{fancy_scientific}
\alias{fancy2_scientific}
\title{String of current time}
\usage{
fancy_scientific(l)

fancy2_scientific(plot, digits = 2)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
