% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publ.R
\name{theme_publ}
\alias{theme_publ}
\title{String of current time}
\usage{
theme_publ(base_size = 12, base_family = "sans", version = 1)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
