% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TJplots.R
\docType{package}
\name{TJplots}
\alias{TJplots}
\alias{TJplots-package}
\title{TJplots: a package with private utility functions for plotting}
\description{
Provide the basic easy-of-use functions for everyday use
}
\section{Utlity functions}{

tjplots
}

